fix: address Copilot review round 3
- Fix stale maxRetries reference in transient HTTP error branch (line 79)
  that bypassed normalization, allowing fractional values to cause the
  function to fall through and return undefined
- Gate network error retries with isTransientError so non-transient fetch
  exceptions (invalid URL, aborts) are thrown immediately without retry
- Merge request.test.js into request.spec.js to follow the project's
  .spec.js naming convention and centralize tests for the request module
- Add test for non-transient fetch error (verifies no retry on plain Error)
